requirements.txt
asgiref==3.8.1
boto3==1.36.5
botocore==1.36.5
crispy-bootstrap5==2024.10
dj-database-url==0.5.0
Django==4.2.16
django-allauth==65.3.1
django-countries==7.6.1
django-crispy-forms==2.3
django-storages==1.14.4
gunicorn==23.0.0
jmespath==1.0.1
pillow==11.1.0
psycopg2-binary==2.9.10
s3transfer==0.11.2
setuptools==75.8.0
sqlparse==0.5.3
stripe==11.4.1